Extract shared crew roster components into CrewRoster module
Move Member type, deriveStatus, RolePill, PlaceholderAvatar,
InvitedAvatar, CrewRow, and CrewSection out of CrewTab into a shared
CrewRoster module so the idea zone can reuse the exact same roster
building blocks. No behavior change — CrewTab now imports what it
previously defined inline.
